+++ src/main/java/com/ankurdave/part/IterCallback.java
@@ -1,5 +1,5 @@
 package com.ankurdave.part;
 
 public interface IterCallback {
-    void apply(final byte[] key, Object value);
+    void apply(final int[] key, Object value);
 }
